Move the code to the folder that matches the current pattern for the central repo
