--- Azure.Configuration.Test/Azure.Configuration.Test.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk" ToolsVersion="15.0">
-  <PropertyGroup>
-    <IsPackable>false</IsPackable>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-    <LangVersion>7.2</LangVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.1" />
-    <PackageReference Include="nunit" Version="$(NUnitPackageVersion)" />
-    <PackageReference Include="NUnit3TestAdapter" Version="$(NUnit3TestAdapterPackageVersion)" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
-    <PackageReference Include="System.Memory" Version="4.5.1" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\Core\Azure.Core\Azure.Core.csproj" />
-    <ProjectReference Include="..\Azure.Configuration\Azure.Configuration.csproj" />
-    <ProjectReference Include="..\..\Core\Azure.Core.Test\Azure.Core.Test.csproj" />
-  </ItemGroup>
-</Project>
